+++ components/canvas_traits/Cargo.toml
@@ -14,6 +14,7 @@ path = "lib.rs"
 webgl_backtrace = []
 
 [dependencies]
+crossbeam-channel = "0.3"
 cssparser = "0.27"
 euclid = "0.20"
 ipc-channel = "0.12"
