+++ components/canvas_traits/lib.rs
@@ -6,6 +6,10 @@
 #![crate_type = "rlib"]
 #![deny(unsafe_code)]
 
+use crate::canvas::CanvasId;
+use crossbeam_channel::Sender;
+use euclid::default::Size2D;
+
 #[macro_use]
 extern crate lazy_static;
 #[macro_use]
@@ -17,3 +21,13 @@ pub mod canvas;
 #[macro_use]
 pub mod webgl;
 mod webgl_channel;
+
+pub enum ConstellationCanvasMsg {
+    Create {
+        id_sender: Sender<CanvasId>,
+        size: Size2D<u64>,
+        webrender_sender: webrender_api::RenderApiSender,
+        antialias: bool,
+    },
+    Exit,
+}
